requirements.txt
alembic==0.9.8
Flask==0.12.2
Flask_Script==2.0.6
Flask_JWT_Extended==3.7.2
SQLAlchemy==1.2.5
Flask_Cors==3.0.3
Flask_Bcrypt==0.7.1
Flask_RESTful==0.3.6
Flask_Migrate==2.1.1
Flask_SQLAlchemy==2.3.2
PyJWT==1.6.0
